requirements.txt
bs4
lxml
GitPython